Problem: rate limiting doesn't work because zloop was never started Solution: It's not possible to simple start the zloop as it takes over the thread control. Thus the zpoller which listened on the pipe has been removed and the task of listening to the pipe has been given to the zloop. Now the zloop can safely take over control from the thread. The zloop will return control if it is interrupted or the terminate command is send.
git-svn-id: https://svn.eckelmann.group/svn/kgl/LDS/Common/zwebrap/trunk@15255 8cdf5627-b312-f249-8d72-5cdd174cf025

+++ src/zwr_response.h
@@ -42,12 +42,12 @@ ZWEBRAP_EXPORT int
    zwr_response_rate_remaining (zwr_response_t *self);
 
 //  Get the rate reset
-ZWEBRAP_EXPORT size_t
+ZWEBRAP_EXPORT int64_t
 zwr_response_rate_reset (zwr_response_t *self);
 
 //  Set parameters for rate limit
 ZWEBRAP_EXPORT void
-zwr_response_set_ratelimit (zwr_response_t *self, int limit, int remaining, size_t reset);
+zwr_response_set_ratelimit (zwr_response_t *self, int limit, int remaining, int64_t reset);
 
 //  Print properties of object
 ZWEBRAP_EXPORT void
